daiy moves as work progresses, Robyne loads, stubs built, pages built, code moved
