+++ README.md
@@ -2,25 +2,8 @@
 
 This project was generated with [Angular CLI](https://github.com/angular/angular-cli) version 15.0.4.
 
-## Development server
 
-Run `ng serve` for a dev server. Navigate to `http://localhost:4200/`. The application will automatically reload if you change any of the source files.
 
-## Code scaffolding
-
-Run `ng generate component component-name` to generate a new component. You can also use `ng generate directive|pipe|service|class|guard|interface|enum|module`.
-
-## Build
-
-Run `ng build` to build the project. The build artifacts will be stored in the `dist/` directory.
-
-## Running unit tests
-
-Run `ng test` to execute the unit tests via [Karma](https://karma-runner.github.io).
-
-## Running end-to-end tests
-
-Run `ng e2e` to execute the end-to-end tests via a platform of your choice. To use this command, you need to first add a package that implements end-to-end testing capabilities.
 
 ## Further help
 
@@ -37,45 +20,15 @@ ng update @angular/core@16 @angular/cli@16
 
 
 ## Robynes adding list for Jay to build
-01b-guest-welcome
-01c3-guest-social-media
-01a-guest-sign-in
-01a-guest-sign-up
+- 01b-guest-welcome
+- 01c3-guest-social-media
+- 01a-guest-sign-in
+- 01a-guest-sign-up
 
 
 ## Michael's moving out list
-* a01c2-guest-about & image
-* a02a-update-your-password ~ had no starting image
-* a04a-hamburger-nav ~ had no starting image
-* b04a-vendor-about-landing-display & image
-* guest-profile ~ had no starting image
+https://dillionmegida.com/p/delete-outdated-branches/
 
-Hello Michael
-
-
-TODO: Michael
-Here is the list of updated image file names:
-a01a-guest-sign-in
-a01a-guest-sign-up
-a01b-guest-welcome
-a01b-guest-welcome-back
-a01c2-guest-about
-a01c3-guest-social-media
-a01c4-guest-web-links
-a02a-update-your-password
-a02a2-update-your-password
-a02a3-update-your-password
-a04a-hamburger-nav
-a04a2-ham-nav-manage-account
-a04a3-ham-nav-manage-alias
-a05a-guest-profile-landing
-a05b-guest-public-landing
-b04a-vendor-about
-b04a-vendor-about-landing-display
-b04b-vendor-bio-landing-display
-b04c-vendor-about
-b04d-vendor-Media-kit -display
-c05c-event-landing
 
 
 This component (a01d1-guest-home) image name has been updated to a05b-guest-public-landing
@@ -94,95 +47,103 @@ These screens are in the works --
 d09a-quest-badges   through   e08b-event-schedule
 
 
-These are all of the components that have updated images as of this session:
-a01a-guest-sign-in
-a01a-guest-sign-up
-a01b-guest-welcome
-a01b-guest-welcome-back
-a01c3-guest-social-media
-a01d1-guest-home
-a01d2-guest-display
-b04a-vendor-bio-about
-b04b-vendor-bio-landing-display
-b04c-vendor-media-build
-b04d-vendor-media-display
-b04e-vendor-shop-display
+These are all of the components that have updated images as of this session 5/21:
+- a01a-guest-sign-in
+- a01a-guest-sign-up
+- a01b-guest-welcome
+- a01b-guest-welcome-back
+- a01c3-guest-social-media
+- a01d1-guest-home
+- a01d2-guest-display
+- b04a-vendor-bio-about
+- b04b-vendor-bio-landing-display
+- b04c-vendor-media-build
+- b04d-vendor-media-display
+- b04e-vendor-shop-display
 
 
 
 TODO:Jay
-Round 3 (fri May 17)
-a01a-guest-sign-in
-a01a-guest-sign-up
-a01b-guest-welcome
-a01c3-guest-social-media
+
+~~Round 3 (fri May 17)~~ **DONE**
+a01a-guest-sign-in *moved*
+a01a-guest-sign-up *moved*
+a01b-guest-welcome *moved*
+a01c3-guest-social-media *moved*
 
 
-Round 4 (sat May 18)
-a01b-guest-welcome-back
-a02a2-update-your-password
-a02a3-update-your-password
-a04a2-ham-nav-manage-account
+~~Round 4 (sat May 18)~~ **DONE**
+a01b-guest-welcome-back *moved*
+a02a2-update-your-password *moved*
+a02a3-update-your-password ***??handled in biocode??***
+a04a2-ham-nav-manage-account   *moved*
 
 
-Round 5 (sun May 19)
-a05a-guest-profile-landing
-b04d-vendor-media-kit-display
-b04e-vendor-shop-displayb04b-vendor-bio-landing-display 
+Round 5 (sun May 19)  **Almost DONE**
+- a05a-guest-profile-landing  *moved*
+- b04d-vendor-media-kit-display *moved*
+- b04e-vendor-shop-display *moved*
+- b04b-vendor-bio-landing-display   **ready to be coded**
 
 
-Round 6 (mon May 20)
-c05c-event-landing
-b04c-vendor-about
-d09a-quest-badges
-d09b-quest-stats 
+Round 6 (mon May 20) **Half DONE**
+- c05c-event-landing  **ready to be coded**
+- b04a-vendor-bio-about  **ready to be coded**
+- d09a-quest-badges *moved*
+- d09b-quest-stats  *moved*
 
 
 Round 7 (tue May 21)
-d09d-quest-rewards
-a05b-guest-taking-qr
-a05c-guest-connection-indicator
-a06a-guest-connecions-landing
+- d09d-quest-rewards  *moved*
+- a05b-guest-taking-qr   **ready to be coded**
+- a05c-guest-connection-indicator  **ready to be coded**
+- a06a-guest-connecions-landing  **ready to be coded**
 
 
 Round 8 (wed May 22)
-a06a1-guest-connecions-landing-swipe-left
-a06a2-guest-connecions-landing-swipe-right
-a06a3-guest-connecions-landing-scroll
-a06b-guest-connecions-landing-search
+- a06a1-guest-connecions-landing-swipe-left  **ready to be coded**
+- a06a2-guest-connecions-landing-swipe-right  **ready to be coded**
+- a06a3-guest-connecions-landing-scroll **ready to be coded**
+- a06b-guest-connecions-landing-search **ready to be coded**
 
 
-Round 9 (thu May 23)
-a06c-guest-requests-landing
-a06c1-guest-requests-landing-swipe-left
-A06c2-guest-requests-landing-swipe-right
-TBD
+Round 9 (thu May 23)  **ready to be coded**
+- a06c-guest-requests-landing  **ready to be coded**
+- a06c1-guest-requests-landing-swipe-left  see differnt views in html
+- a06c2-guest-requests-landing-swipe-right   see differnt views in html
+- d09e-quests-redeemed   **ready to be coded**
 
 
-TODO: Michael
 
-a05b-guest-taking-qr
-a05c-guest-connection-indicator
-a06a-guest-connecions-landing
-a06a1-guest-connecions-landing-swipe-left
-a06a2-guest-connecions-landing-swipe-right
-a06a3-guest-connecions-landing-scroll
-a06b-guest-connecions-landing-search
-a06c-guest-requests-landing
-a06c1-guest-requests-landing-swipe-left
-A06c2-guest-requests-landing-swipe-right
+TODO:Robyne (5/22)
+- e08b-event-schedule ***waiting to be drawn***
 
 
-TODO:Robyne
+
+TODO: Michael 5/21  *build & check components*
+~~a05b-guest-taking-qr~~
+~~a05c-guest-connection-indicator~~
+~~a06a-guest-connecions-landing~~
+~~a06a1-guest-connecions-landing-swipe-left~~
+~~a06a2-guest-connecions-landing-swipe-right~~
+~~a06a3-guest-connecions-landing-scroll~~
+~~a06b-guest-connecions-landing-search~~
+~~a06c-guest-requests-landing~~
+~~a06c1-guest-requests-landing-swipe-left~~
+~~a06c2-guest-requests-landing-swipe-right~~
+
+TODO:Robyne  5/21
 file organization and clean up ASAP!
 work on additional screens to show how to view other guest profiles from the connections and requests screens and show a communication thread
 work on translating Kevin's diagram
 work on questing sequence
 
-TODO: Michael
+TODO: Michael  5/22
 Screen for b04f-vendor-store-connect doesn't exist in figma. Not sure if this is a typo or if you're thinking we need an additional screen?
 same as above for b04g-vendor-store-connect-codes
+
 
 These component names will need to be updated to match the image names
 d09c-quest-qr has should be changed to a05b-guest-taking-qr
 d09e-quests-redeemed should be changed to d09e-quests-redeemed
+
