+++ src/app/a05b-guest-taking-qr/a05b-guest-taking-qr.component.html
@@ -1,1 +1,2 @@
 <p>a05b-guest-taking-qr works!</p>
+<img src="/src/assets/images/design-photos/a05b-guest-taking-qr.png" />
