+++ src/app/a05c-guest-connection-indicator/a05c-guest-connection-indicator.component.html
@@ -1,1 +1,3 @@
 <p>a05c-guest-connection-indicator works!</p>
+
+<img src="assets/images/a05c-guest-connection-indicator.png" alt="Guest Connection Indicator" />
