+++ src/app/a06c-guest-requests-landing/a06c-guest-requests-landing.component.html
@@ -1,2 +1,5 @@
 <p>a06c-guest-requests-landing works!</p>
-A06c2-guest-requests-landing-swipe-right
+
+<img src="assets/images/a06c-guest-requests-landing" alt="Guest Requests Landing" />
+a06c1-guest-requests-landing-swipe-left
+a06c2-guest-requests-landing-swipe-right
