--- src/app/b04d-vendor-media-display/b04d-vendor-media-display.component.html
@@ -1,2 +1,0 @@
-<p>b04d-vendor-media-display works!</p>
-<img src="/src/assets/images/design-photos/b04d-vendor-Media-kit -display.png" alt="Vendor Media Display"/>
