--- src/app/b04d-vendor-media-display/b04d-vendor-media-display.component.ts
@@ -1,10 +1,0 @@
-import { Component } from '@angular/core';
-
-@Component({
-  selector: 'app-b04d-vendor-media-display',
-  templateUrl: './b04d-vendor-media-display.component.html',
-  styleUrls: ['./b04d-vendor-media-display.component.css']
-})
-export class B04dVendorMediaDisplayComponent {
-
-}
