+++ src/app/built/a01b-guest-welcome-back/a01b-guest-welcome-back.component.html
@@ -1,8 +1,7 @@
 
 <div class="pt-44">
     <h1 class="font-extrabold text-4xl text-center px-20 uppercase text-gray-700 ">
-        Welcome,
-        John!
+        Welcome, John!
     </h1>
 </div>
