+++ src/app/built/b04e-vendor-shop-display/b04e-vendor-shop-display.component.ts
