+++ src/app/c05c-event-landing/c05c-event-landing.component.html
@@ -1,0 +1,2 @@
+<p>c05c-event-landing works!</p>
+<img src="/src/assets/images/design-photos/c05c-event-landing.png" />
