+++ src/app/c05c-event-landing/c05c-event-landing.component.ts
@@ -1,0 +1,10 @@
+import { Component } from '@angular/core';
+
+@Component({
+  selector: 'app-c05c-event-landing',
+  templateUrl: './c05c-event-landing.component.html',
+  styleUrls: ['./c05c-event-landing.component.css']
+})
+export class C05cEventLandingComponent {
+
+}
