+++ src/app/e08b-event-schedule/e08b-event-schedule.component.html
@@ -1,1 +1,2 @@
 <p>e08b-event-schedule works!</p>
+<img src="/src/assets/images/design-photos/------.png" />
