--- src/app/guest-connection-indicator/guest-connection-indicator.component.html
@@ -1,1 +1,0 @@
-<p>guest-connection-indicator works!</p>
