--- src/app/guest-connection-indicator/guest-connection-indicator.component.ts
@@ -1,10 +1,0 @@
-import { Component } from '@angular/core';
-
-@Component({
-  selector: 'app-guest-connection-indicator',
-  templateUrl: './guest-connection-indicator.component.html',
-  styleUrls: ['./guest-connection-indicator.component.css']
-})
-export class GuestConnectionIndicatorComponent {
-
-}
