--- src/app/guest-taking-qr/guest-taking-qr.component.html
@@ -1,1 +1,0 @@
-<p>guest-taking-qr works!</p>
