--- src/app/guest-taking-qr/guest-taking-qr.component.ts
@@ -1,10 +1,0 @@
-import { Component } from '@angular/core';
-
-@Component({
-  selector: 'app-guest-taking-qr',
-  templateUrl: './guest-taking-qr.component.html',
-  styleUrls: ['./guest-taking-qr.component.css']
-})
-export class GuestTakingQrComponent {
-
-}
