+++ src/assets/images/design-photos/b04e-vendor-shop-display.png
